refactor : 해당 도메인으로 exception이동
