+++ src/main/java/kitchenpos/order/application/OrderService.java
@@ -6,12 +6,12 @@ import kitchenpos.order.dto.OrderRequest;
 import kitchenpos.order.dto.OrderResponse;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuRepository;
-import kitchenpos.common.exception.NotExistMenuException;
-import kitchenpos.common.exception.NotExistOrderException;
-import kitchenpos.common.exception.UnableCreateOrderException;
+import kitchenpos.menu.exception.NotExistMenuException;
+import kitchenpos.order.exception.NotExistOrderException;
+import kitchenpos.order.exception.UnableCreateOrderException;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.OrderTableRepository;
-import kitchenpos.common.exception.NotExistOrderTableException;
+import kitchenpos.table.exception.NotExistOrderTableException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.CollectionUtils;
