+++ src/main/java/kitchenpos/product/exception/NotExistProductException.java
@@ -1,4 +1,4 @@
-package kitchenpos.common.exception;
+package kitchenpos.product.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
@@ -6,11 +6,11 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import java.util.NoSuchElementException;
 
 @ResponseStatus(HttpStatus.BAD_REQUEST)
-public class NotExistOrderException extends NoSuchElementException {
-    public NotExistOrderException() {
+public class NotExistProductException extends NoSuchElementException {
+    public NotExistProductException() {
     }
 
-    public NotExistOrderException(String s) {
+    public NotExistProductException(String s) {
         super(s);
     }
 }
