+++ src/main/java/kitchenpos/table/application/TableService.java
@@ -7,7 +7,7 @@ import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.OrderTableRepository;
 import kitchenpos.table.dto.OrderTableRequest;
 import kitchenpos.table.dto.OrderTableResponse;
-import kitchenpos.common.exception.NotExistOrderTableException;
+import kitchenpos.table.exception.NotExistOrderTableException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
