+++ src/main/java/kitchenpos/table/exception/NotExistTableException.java
@@ -1,4 +1,4 @@
-package kitchenpos.common.exception;
+package kitchenpos.table.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
@@ -6,11 +6,11 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import java.util.NoSuchElementException;
 
 @ResponseStatus(HttpStatus.BAD_REQUEST)
-public class NotExistMenuException extends NoSuchElementException {
-    public NotExistMenuException() {
+public class NotExistTableException extends NoSuchElementException {
+    public NotExistTableException() {
     }
 
-    public NotExistMenuException(String s) {
+    public NotExistTableException(String s) {
         super(s);
     }
 }
