+++ src/main/java/kitchenpos/table/exception/UnableChangeNumberOfGuestsException.java
@@ -1,4 +1,4 @@
-package kitchenpos.common.exception;
+package kitchenpos.table.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
